refactored code to follow best practices in spring

+++ contracts/src/main/java/clm/demo/dto/requests/GenContractRequest.java
@@ -5,11 +5,14 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.NotEmpty;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
 /**
  * @param mappings LABEL - VALUE MAPPINGS
+ * @param value    contract monetary value; {@code BigDecimal} used instead of
+ *                 {@code Double} to avoid floating-point rounding issues in financial contexts
  */
 public record GenContractRequest(@NotNull(message = "Template ID is required") Long templateId,
                                  @NotNull(message = "User ID is required") Long userId,
@@ -18,6 +21,6 @@ public record GenContractRequest(@NotNull(message = "Template ID is required") L
                                  @NotNull(message = "Start date is required") LocalDate startDate,
                                  @NotNull(message = "End date is required") LocalDate endDate,
                                  @NotNull(message = "Field mappings map is required") Map<String, String> mappings,
-                                 Double value, String notes) {
+                                 BigDecimal value, String notes) {
 
 }
